(maint) Allow bundler to resolve lockfile via local gems
Previously, when we generated the Gemfile.lock, we passed in the
PUPPET_GEM_VERSION env variable to set puppet version to the
latest_puppet as defined in the matrix. Instead, we should be using
the matrix to install those versions of puppet in the gem cache, then
allowing bundler to resolve the versions on its own.