(PDK-1168) Add guard in PS Module to not run on ISE
Previously users could run PDK within the Windows PowerShell ISE application and
it would hang while waiting for user input.  This is a known issue with the ISE
dating back to 2009 [1].  This commit adds a guard on the PowerShell Host called
'Windows PowerShell ISE Host' and throws a helpful error message to the user and
terminates early, not running PDK and not hanging.

[1] https://devblogs.microsoft.com/powershell/console-application-non-support-in-the-ise/