(PDK-1370) Create tarballs of ruby environment for Windows
Previously the PDK build process created 50,000+ files which causes issues with
MSI installation and uninstallation times.  This commit is the start of a
process to decrease install time by generating less files.

This commit:
* Adds a new component called pdk-create-ruby-tarballs which is responsible for
  taking a PDK installation files and compressing the ruby environments and
  their respective gem and puppet gem caches.  Note that the ruby environment
  which PDK runs in, at this time Ruby 2.4 is not compressed as it is required
  for the decompression step.
* The new pdk-create-ruby-tarballs component will only run on Windows builds.
  Later commits may remove this restriction.
* Adds a new file the to installation media called `extract_all.rb`. This small
  script is responsible for extracting the tarballs back to their original
  locations.  Later commits will use this ruby file, during installation time,
  to "re-hydrate" the entire PDK ruby and gem files.

For the moment this commit is only responsible for creating the tarballs and
adding the extraction script.

Note - Compresses files are removed from the installation source media which
resulted in over 30,000 files being moved into the compressed tarballs.