(PDK-1468) Refactor executing ruby in MSI Custom Actions
Previously the ruby custom action assumed that the PDK install directory would
always have a shortname, however it is possible to disable 8.3 (short name)
generation.  This is common in virtualised environments or Terminal Servers.

This commit:
* Changes the call by wrapping the windows paths in double quotes to
ensure the path is interpreted correctly by cmd.exe.
* Due to strange double quoting behaviour in wshShell.Run we instead need to use
an intermediate batch file with the actual command and then call that.
* The intermediate batch file adds the EXIT command at the end to ensure that
even if the batch file contents are modified during its lifetime, any malicious
code will be ignored