(PDK-1547) Set SSL Env vars before running PDK in PowerShell Module
Previously the the PDK PowerShell module was susceptible to external environment
variable changes once the module was loaded.  For example run PDK, then Bolt and
then PDK again, it then fails with a Ruby error.  This was due to the module
setting Environment Variables only once during the module loading process.

This commit:
* Changes some of the variables to be module scoped ($Script) instead of
  environment, as it's not needed.
* Modifies the SSL based env. vars prior to invoking PDK every time, instead of
  only once.