(MAINT) Refactor `puppet-forge-api` component into `puppet-versions`
The `puppet-forge-api` component originally pulled in multiple files
from the Forge API codebase (which is a private repo) but at this point
it was only pulling in the PE version mapping file which is now
available from a public API endpoint that we can just pull from at build
time.

This PR refactors the component into a "file" based component that
queries the Rubygems API to gather a list of Puppet gem versions and
installs them into the appropriate gem caches.