(CAT-1299) Expand documentation
Prior to this commit, the Readme.md file in this repository barely had
a short description and usage block with instructions.

This commit aims to enforce the standard Puppet documentation guidelines
by formatting the document properly and to expand a bit the existing
documentation with relevant information about the tool.